Move resourceLoader logic to find maps to AvialableGAmesList
Consolidates the operations and management of maps-on-disk
to one class.
